【ParentDeptId -> ParentDeptName】上级部门id转换为上级部门名称【4.5】

+++ src/main/java/com/aioveu/boot/aioveuDepartment/model/entity/AioveuDepartment.java
@@ -67,7 +67,7 @@ public class AioveuDepartment {
     /**
      * 上级部门ID
      */
-    private Integer parentDeptId;
+    private Long parentDeptId;
     /**
      * 部门经理
      */
